+++ test/MultipleChoiceScene.test.js
@@ -1,9 +1,10 @@
 import React from 'react';
-import { configure, shallow } from 'enzyme';
+import { configure, shallow, mount } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 
 import { MultipleChoiceScene } from '../src/MultipleChoiceScene';
 import { ChoiceList } from '../src/ChoiceList';
+import { CommandLink } from '../src/CommandLink';
 
 configure({ adapter: new Adapter() });
 
@@ -15,9 +16,10 @@ describe('<MultipleChoiceScene />', () => {
                 prompt: "Choose"
             }
         }
-        let scene = shallow(<MultipleChoiceScene {...props} />);
+        let scene = mount(<MultipleChoiceScene {...props} />);
         expect(scene.find(ChoiceList).length).toBe(1);
-    })
+        expect(scene.find(CommandLink).length).toBe(2);
+    });
 
     it('renders children', () => {
         let scene = shallow(<MultipleChoiceScene><img /></MultipleChoiceScene>);